try to launch kernel
