+++ test/executor_test.cc
@@ -25,7 +25,7 @@ int main() {
   mscclpp::ExecutionPlan plan(MSCCLPP_ROOT_PATH + "/test/execution-files/allreduce.json");
   std::shared_ptr<char> sendbuff = mscclpp::allocExtSharedCuda<char>(1024);
   std::shared_ptr<char> recvbuff = mscclpp::allocExtSharedCuda<char>(1024);
-  executor->execute(rank, sendbuff.get(), recvbuff.get(), 1024, 1024, plan);
+  executor->execute(rank, sendbuff.get(), recvbuff.get(), 1024, 1024, 512, plan);
 
   MPI_Finalize();
   return 0;
